Move registry module.
