+++ api/admin/controller/discovery_service_library_registrations.py
@@ -5,7 +5,7 @@ from flask import Response
 from flask_babel import lazy_gettext as _
 
 from api.admin.problem_details import *
-from api.registry import Registration, RemoteRegistry
+from api.registration.registry import Registration, RemoteRegistry
 from core.model import ExternalIntegration, Library, get_one
 from core.util.http import HTTP
 from core.util.problem_detail import ProblemDetail
