+++ bin/configuration/register_library
@@ -6,6 +6,6 @@ import sys
 bin_dir = os.path.split(__file__)[0]
 package_dir = os.path.join(bin_dir, "..", "..")
 sys.path.append(os.path.abspath(package_dir))
-from api.registry import LibraryRegistrationScript
+from api.registration.registry import LibraryRegistrationScript
 
 LibraryRegistrationScript().run()
