+++ tests/admin/test_geographic_validator.py
@@ -6,11 +6,10 @@ import urllib.request
 import pypostalcode
 import uszipcode
 
-from api.admin.controller.library_settings import LibrarySettingsController
 from api.admin.geographic_validator import GeographicValidator
 from api.admin.problem_details import *
 from api.config import Configuration
-from api.registry import RemoteRegistry
+from api.registration.registry import RemoteRegistry
 from core.model import ExternalIntegration, create
 from core.testing import MockRequestsResponse
 from tests.admin.controller.test_controller import SettingsControllerTest
